+++ apps/server/src/binaries/binaries.module.ts
@@ -1,13 +1,14 @@
 import { Module } from '@nestjs/common';
 import { BinariesController } from './binaries.controller';
 import { BinariesService } from './binaries.service';
+import { BinariesInstallerService } from './installer.service';
 import { BinaryResourcesService } from './binary-resources.service';
 import { SystemModule } from '../system/system.module';
 
 @Module({
   imports: [SystemModule],
   controllers: [BinariesController],
-  providers: [BinariesService, BinaryResourcesService],
+  providers: [BinariesService, BinaryResourcesService, BinariesInstallerService],
   exports: [BinariesService, BinaryResourcesService]
 })
 export class BinariesModule {}